make distributable docker image script.

+++ sh/devcli.sh
@@ -2,6 +2,21 @@
 
 base_yml=docker-compose.python3.base.yml
 
+while getopts "hb" arg; do
+    case $arg in
+      h)
+        echo ""
+        echo "usage: [-h] [-b build but not run]"
+        echo ""
+        exit
+        ;;
+      b)
+        build_only=true
+        ;;
+    esac
+done
+
+
 if [ -z $PYTHON_VER ]
 then
     base_yml=docker-compose.python3.base.yml
@@ -10,7 +25,10 @@ else
 fi
 
 set -x
-docker-compose -f $base_yml build
+docker-compose -f ${base_yml} build
 docker-compose -f docker-compose.dev.yml build
-docker-compose -f docker-compose.dev.yml run --rm --name dev-cli cron-serv-dev-cli
+if [ -z $build_only ]
+then
+    docker-compose -f docker-compose.dev.yml run --rm --name dev-cli cron-serv-dev-cli
+fi
 exit
